Update "new credits check" page to once again show a confirmation dialogue box on form submission
