+++ mtp_noms_ops/assets-src/__old__/javascripts/main.js
@@ -2,7 +2,6 @@
   'use strict';
 
   // noms-ops
-  require('confirm-checked').ConfirmChecked.init();
   require('choose-prisons').ChoosePrisons.init();
   require('toggle-text-field-on-change').ToggleTextFieldOnChange.init();
 }());
