+++ mtp_noms_ops/templates/security/review.html
@@ -24,7 +24,7 @@
 </p>
 
 {% if credits %}
-  <form method="post" class="mtp-review js-ConfirmChecked mtp-form-analytics">
+  <form method="post" class="mtp-review mtp-form-analytics">
     {% csrf_token %}
 
     <p class="govuk-!-display-none-print">
